server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "django==5.1.4",
    "django-cors-headers==4.6.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "idna==3.10",
    "pillow==11.1.0",
    "pyjwt==2.10.1",
    "python-decouple==3.8",
    "requests==2.32.3",
    "sqlparse==0.5.3",
    "stripe==11.4.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.3.0",
]
